requirements_local.txt
packaging
safetensors
requests
torch
torchvision
python-multipart
fastapi
uvicorn
jinja2
aiofiles
pillow
numpy
